requirements.txt
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
flasgger==0.9.7.1
Flask==3.1.0
Flask-JWT-Extended==4.7.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.0.2
packaging==24.2
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.0
six==1.16.0
urllib3==2.2.3
Werkzeug==3.1.3
